+++ include/dcc/address.hpp
@@ -28,8 +28,8 @@ struct Address {
   }
 
   friend constexpr bool operator==(Address const& lhs, Address const& rhs) {
-    return (lhs.type == Short || lhs.type == Long) &&
-               (rhs.type == Short || rhs.type == Long)
+    return (lhs.type == BasicLoco || lhs.type == ExtendedLoco) &&
+               (rhs.type == BasicLoco || rhs.type == ExtendedLoco)
              ? lhs.value == rhs.value
              : lhs.value == rhs.value && lhs.type == rhs.type;
   }
@@ -40,15 +40,16 @@ struct Address {
   value_type value{};
 
   enum : uint8_t {
-    UnknownService, ///< Address is unknown or service (=no address)
-    Broadcast,      ///< Address is broadcast
-    Short,          ///< Address is short
-    Accessory,      ///< Address is accessory decoder
-    Long,           ///< Address is long
-    Reserved,       ///< Address is reserved
-    DataTransfer,   ///< Address for data transfer
-    AutomaticLogon, ///< Address is automatic logon
-    IdleSystem      ///< Address for system commands
+    UnknownService,    ///< Unknown or service (=no address)
+    Broadcast,         ///< Broadcast
+    BasicLoco,         ///< Basic loco ()
+    BasicAccessory,    ///< Basic accessory
+    ExtendedAccessory, ///< Extended accessory
+    ExtendedLoco,      ///< Extended loco
+    Reserved,          ///< Reserved
+    DataTransfer,      ///< Data transfer
+    AutomaticLogon,    ///< Automatic logon
+    IdleSystem         ///< System commands
   } type{};
 
   bool reversed{}; /// Direction reversed
@@ -64,20 +65,22 @@ constexpr Address decode_address(InputIt first) {
   // 0
   if (*first == 0u) return {*first, Address::Broadcast};
   // 1-127
-  else if (*first <= 127u) return {*first, Address::Short};
+  else if (*first <= 127u) return {*first, Address::BasicLoco};
   // 128-191
   else if (*first <= 191u) {
     auto const a7_2{(*first++ & 0b0011'1111u) << 2u};
     auto const a10_8{(*first & 0b0111'0000u) << 4u};
     auto const a1_0{(*first & 0b0000'0110u) >> 1u};
     return {static_cast<Address::value_type>(a10_8 | a7_2 | a1_0),
-            Address::Accessory};
+            *first & 0b1000'0000u ? Address::BasicAccessory
+                                  : Address::ExtendedAccessory};
   }
   // 192-231
   else if (*first <= 231u) {
     auto const a13_8{(*first++ & 0b0011'1111u) << 8u};
     auto const a7_0{*first};
-    return {static_cast<Address::value_type>(a13_8 | a7_0), Address::Long};
+    return {static_cast<Address::value_type>(a13_8 | a7_0),
+            Address::ExtendedLoco};
   }
   // 232-252
   else if (*first <= 252u)
@@ -117,14 +120,22 @@ constexpr OutputIt encode_address(Address addr, OutputIt first) {
   switch (addr.type) {
     case Address::UnknownService: break;
     case Address::Broadcast: *first++ = 0u; break;
-    case Address::Short: *first++ = static_cast<uint8_t>(addr); break;
-    case Address::Accessory:
+    case Address::BasicLoco: *first++ = static_cast<uint8_t>(addr); break;
+    case Address::BasicAccessory:
       *first++ =
         static_cast<uint8_t>(0b1000'0000u | ((addr & 0b1111'1100u) >> 2u));
-      *first++ = static_cast<uint8_t>((addr & 0b0111'0000'0000u) >> 4u |
-                                      (addr & 0b0000'0011u) << 1u);
+      *first++ =
+        static_cast<uint8_t>(0b1000'0000u | (addr & 0b0111'0000'0000u) >> 4u |
+                             (addr & 0b0000'0011u) << 1u);
       break;
-    case Address::Long:
+    case Address::ExtendedAccessory:
+      *first++ =
+        static_cast<uint8_t>(0b1000'0000u | ((addr & 0b1111'1100u) >> 2u));
+      *first++ =
+        static_cast<uint8_t>((addr & 0b0111'0000'0000u) >> 4u |
+                             (addr & 0b0000'0011u) << 1u | 0b0000'0001u);
+      break;
+    case Address::ExtendedLoco:
       *first++ = static_cast<uint8_t>(0b1100'0000u | addr >> 8u);
       *first++ = static_cast<uint8_t>(addr);
       break;