+++ tests/tx/capacity.cpp
@@ -1,0 +1,3 @@
+#include "tx_test.hpp"
+
+TEST_F(TxTest, capacity) { EXPECT_EQ(_mock.capacity(), DCC_TX_DEQUE_SIZE); }