+++ tests/tx/tx_test.cpp
@@ -1,0 +1,5 @@
+#include "tx_test.hpp"
+
+TxTest::TxTest() {}
+
+TxTest::~TxTest() {}