+++ tests/tx/tx_test.hpp
@@ -1,0 +1,15 @@
+#pragma once
+
+#include <gtest/gtest.h>
+#include "tx_mock.hpp"
+
+using namespace ::testing;
+
+// Transmit test fixture
+struct TxTest : ::testing::Test {
+protected:
+  TxTest();
+  virtual ~TxTest();
+
+  NiceMock<TxMock> _mock;
+};